Modifiqué Paises.h/Paises.cpp para usar set en vez de map y realicé algunos ajustes en la precisión de los decimales(no todas las coordenadas tienen los mismos decimales y a la hora de mostrarlo en texto esto perjudica la salida) por lo demas la clase Paises parece funcionar correctamente

+++ rutas_aereas/src/Punto.cpp
@@ -42,8 +42,8 @@ istream & operator>>(istream &is,Punto &punto){
 }
 
 pair<double, double> punto_aCoordenadas(const Punto &punto, const int &totalfilas, const int &totalcolumnas) {
-    double latitud = 90 - (punto.getX() / (totalfilas / 180.0));
-    double longitud = (punto.getY() / (totalcolumnas / 360.0)) - 180;
+    double latitud = 90 - (punto.getX() / ((double)totalfilas / 180.0));
+    double longitud = (punto.getY() / ((double)totalcolumnas / 360.0)) - 180;
     return {latitud, longitud};
 }
